rename convert_tools to plain convert
